# improve palette ordering and encode speed