@jonsneyers 2023-06-08
Further simplified the code a bit to try global RGBA/RGB palette only once, before trying channel palette, as opposed to the previous strategy of first trying it with few max colors and then again after channel palette with more max colors.
This ensures that the palette ordering always makes sense (at least for global palettes). Doing channel palette first and _then_ RGB(A) palette can save some bytes in weird cases (like 16-bit images with ~1000 unique colors so palette can be done but the cost of signaling the palette is not so small), but usually it will not help much or even be counterproductive, and in any case it will make decoding slower (need to undo multiple palettes instead of just one) and makes the palette ordering weird since the channel ranges will be different (and the sorting kind of assumes they're the same).